cron.php is reworked, other worker files had been moved temporarily
