+++ src/Core/Worker.php
@@ -241,7 +241,7 @@ class Worker {
 
 		// The script could be provided as full path or only with the function name
 		if ($include == basename($include)) {
-			$include = "include/".$include.".php";
+			$include = "worker/".$include.".php";
 		}
 
 		if (!validate_include($include)) {
@@ -902,7 +902,7 @@ class Worker {
 		self::add(PRIORITY_HIGH, "spool_post");
 
 		// Run the cron job that calls all other jobs
-		self::add(PRIORITY_MEDIUM, "cron");
+		self::add(PRIORITY_MEDIUM, "Cron");
 
 		// Run the cronhooks job separately from cron for being able to use a different timing
 		self::add(PRIORITY_MEDIUM, "CronHooks");
