+++ worker/create_shadowentry.php
