+++ worker/discover_poco.php
