+++ worker/remove_contact.php
